Revision summary: -Changes in logging system with the ability to rollback to the original console-logging. -logWrite macro changed to log(format,va_args) and logEx(field,format,va_args) -*.hpp <- *.h , \\ <- / -Removed warnings -Type conversion fixes. There's still a lot of work to be done here..Especially in template handling which could affect generated code(if not already.. which is what i believe).
Revision notes (programmers):
To rollback to original logging method that you're used to , open log/logging.h
and replace "#undef DEFAULT_LOGGING" with "#define DEFAULT_LOGGING".

Revision notes (end users):
For maximum performance it is recommended to leave logging disabled.
If you encounter a bug , enable it to provide the call trace.

+++ trunk/nulldc/nullDC/dc/sh4/tmu.cpp
@@ -72,7 +72,7 @@ void UpdateTMUCounts(u32 reg)
 	InterruptPend(tmu_intID[reg],tmu_regs_CR[reg] & tmu_underflow);
 	InterruptMask(tmu_intID[reg],tmu_regs_CR[reg] & tmu_UNIE);
 
-	if (old_mode[reg]==(tmu_regs_CR[reg] & 0x7))
+	if (old_mode[reg]==(u32)(tmu_regs_CR[reg] & 0x7))
 		return;
 	else
 		old_mode[reg]=(tmu_regs_CR[reg] & 0x7);
